refact: reorganize authentication schemas and remove unused mobile trigger component; update imports for clarity
