+++ src/components/modules/auth/hooks/useAuth.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useForm } from "react-hook-form";
-import { authSchema } from "../ui/schema/auth.schema";
+import { authSchema } from "../schema/auth.schema";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
 import { supabase } from "@/lib/supabase";
