--- src/components/modules/landing/hooks/feature-card-animations.hook.ts
@@ -1,12 +1,0 @@
-"use client";
-
-import { useAnimationVariants } from "./animation-variants.hook";
-
-export const useFeatureCardAnimations = (index: number) => {
-  const { cardVariants, iconVariants } = useAnimationVariants();
-
-  return {
-    cardVariants: cardVariants(index),
-    iconVariants: iconVariants(index),
-  };
-};
