--- src/components/modules/landing/hooks/footer-animations.hook.ts
@@ -1,13 +1,0 @@
-"use client";
-
-export const useFooterAnimations = () => {
-  const footerLinks = [
-    { href: "/terms", label: "Terms" },
-    { href: "/privacy", label: "Privacy" },
-    { href: "https://docs.trustlesswork.com", label: "Documentation" },
-  ];
-
-  return {
-    footerLinks,
-  };
-};
