--- src/components/modules/landing/hooks/step-card-animations.hook.ts
@@ -1,12 +1,0 @@
-"use client";
-
-import { useAnimationVariants } from "./animation-variants.hook";
-
-export const useStepCardAnimations = (index: number) => {
-  const { cardVariants, iconVariants } = useAnimationVariants();
-
-  return {
-    cardVariants: cardVariants(index),
-    circleVariants: iconVariants(index),
-  };
-};
