--- src/components/modules/landing/hooks/use-case-card-animations.hook.ts
@@ -1,14 +1,0 @@
-"use client";
-
-import { useAnimationVariants } from "./animation-variants.hook";
-
-export const useUseCaseCardAnimations = (index: number) => {
-  const { cardVariants, listVariants, listItemVariants } =
-    useAnimationVariants();
-
-  return {
-    cardVariants: cardVariants(index),
-    listVariants: listVariants(index),
-    itemVariants: listItemVariants,
-  };
-};
